# openapi.yaml
openapi: 3.0.3
info:
  title: Sharpl API
  description: API specification for Sharpl application
  version: 1.0.0
  contact:
    name: Sharpl Team
    
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:3000/api
    description: Frontend proxy

tags:
  - name: auth
    description: Authentication operations
  - name: users
    description: User management operations
  - name: health
    description: Health check operations

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Check if the server is running
      operationId: getHealth
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ping:
    get:
      tags:
        - health
      summary: Ping endpoint
      description: Simple ping endpoint for testing API connectivity
      operationId: getPing
      responses:
        '200':
          description: Pong response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
    
    post:
      tags:
        - health
      summary: Ping with body
      description: Ping endpoint that echoes back the request body
      operationId: postPing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Pong response with echo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingEchoResponse'
        '400':
          description: Invalid JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/login:
    post:
      tags:
        - auth
      summary: User login
      description: Authenticate user and return token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/register:
    post:
      tags:
        - auth
      summary: User registration
      description: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/logout:
    post:
      tags:
        - auth
      summary: User logout
      description: Logout current user
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  /api/v1/users:
    get:
      tags:
        - users
      summary: Get all users
      description: Retrieve a list of all users
      operationId: getUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'

  /api/v1/users/{id}:
    get:
      tags:
        - users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - email

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: password123
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        message:
          type: string
          example: Login successful
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - message
        - user

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: newuser@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: password123
        name:
          type: string
          example: John Doe
      required:
        - email
        - password

    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
        email:
          type: string
          format: email
          example: newuser@example.com

    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          example: 10

    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: healthy
        message:
          type: string
          example: Server is running

    PingResponse:
      type: object
      properties:
        message:
          type: string
          example: pong
        time:
          type: string
          format: date-time
          example: 2026-01-05T13:45:00Z

    PingEchoResponse:
      type: object
      properties:
        message:
          type: string
          example: pong
        time:
          type: string
          format: date-time
        received:
          type: object
          additionalProperties: true

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation successful

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials